+++ liquid-compiler/src/lib.rs
@@ -10,7 +10,7 @@ mod block;
 mod filter;
 mod filter_chain;
 mod include;
-mod options;
+mod lang;
 mod parser;
 mod registry;
 mod tag;
@@ -20,7 +20,7 @@ pub use block::*;
 pub use filter::*;
 pub use filter_chain::*;
 pub use include::*;
-pub use options::*;
+pub use lang::*;
 pub use parser::*;
 pub use registry::*;
 pub use tag::*;
